ANALISE_GLASSDOOR/requirements.txt
pandas
pyodbc
nltk
scikit-learn
transformers
torch
python-dotenv
openpyxl
bs4
selenium
beautifulsoup4
webdriver_manager
webdriver_manager.chrome
requests